Add item items (#338)

+++ src/SerialLoops.Lib/Items/BackgroundItem.cs
@@ -35,7 +35,7 @@ namespace SerialLoops.Lib.Items
             if (cgEntry is not null)
             {
                 CgName = cgEntry?.Name?.GetSubstitutedString(project);
-                ExtrasShort = cgEntry?.Flag ?? 0;
+                ExtrasShort = (short)(cgEntry?.Flag ?? 0);
                 ExtrasInt = cgEntry?.Unknown04 ?? 0;
             }
             PopulateScriptUses(project.Evt);
